:recycle: Refactor moving a utility methods to other file.
